Make it a help-json command instead of a help flag
Commander's implicit help command takes no options, so `help --json` looked
up a subcommand named `--json` and silently printed root help. Rather than
unregistering that command and re-implementing it, `help-json [command...]`
stands on its own — plain help stays Commander's, and lists `help-json` with
its description so an agent finds it from `explorbot help`.

`HelpCommand` goes back to the TUI stub it was; the tree builder lives in
`HelpJsonCommand`.
